Track current query on DynamicPicker
We need to compare the current query in the file picker to the query
that we recorded on the last completion request. If we don't, we won't
be able to use the arrow keys: using the arrow keys will trigger
another idle timeout, so if we scroll through the options we end up
re-requesting the workspace symbols which resets our position in the
options, making it impossible to select anything but the first result.